+++ CHANGELOG.md
@@ -11,8 +11,14 @@ This project adheres to [Semantic Versioning](http://semver.org).
 
 ### Changed
 
-* `ItemInterface` objects add themselves as children to their parents
-* `ItemInterface` parents verify if they are the parent of their children
+* Renamed `Item` to `WzItem`
+* Renamed `ItemInterface` to `WzItemInterface`
+* Renamed `ItemManager` to `WzItemCollection`
+* Renamed `ItemManagerInterface` to `WzItemCollectionInterface`
+* Renamed `CompleteClassificationXml` to `WzClassification`
+* Renamed `CompleteClassificationXmlFile` to `WzClassificationFile`
+* `WzItemInterface` objects add themselves as children to their parents
+* `WzItemInterface` parents verify if they are the parent of their children
 
 ## 1.0.0-rc1 - 2017-02-05